# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PharmaCare Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /medicines:
    get:
      operationId: listMedicines
      tags: [medicines]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: companyId
          in: query
          schema:
            type: integer
        - name: genericNameId
          in: query
          schema:
            type: integer
        - name: isControlled
          in: query
          schema:
            type: boolean
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of medicines with stock
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicineWithStock"
    post:
      operationId: createMedicine
      tags: [medicines]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicineBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
  /medicines/{id}:
    get:
      operationId: getMedicine
      tags: [medicines]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medicine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicineWithStock"
    patch:
      operationId: updateMedicine
      tags: [medicines]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMedicineBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
    delete:
      operationId: deleteMedicine
      tags: [medicines]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /medicines/{id}/batches:
    get:
      operationId: getMedicineBatches
      tags: [medicines]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Batches for medicine
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Batch"
  /medicines/{id}/alternatives:
    get:
      operationId: getMedicineAlternatives
      tags: [medicines]
      summary: Get medicines with the same generic name
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alternative medicines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicineWithStock"
...
  /purchase-returns:
    get:
      operationId: listPurchaseReturns
      tags: [purchaseReturns]
      parameters:
        - name: supplierId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of purchase returns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseReturn"
    post:
      operationId: createPurchaseReturn
      tags: [purchaseReturns]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseReturnBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseReturn"
  /purchase-returns/{id}:
    get:
      operationId: getPurchaseReturn
      tags: [purchaseReturns]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase return with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseReturnWithItems"

  # ─── SALES / POS ────────────────────────────────────────────────────────────
...
  /missed-sales:
    get:
      operationId: listMissedSales
      tags: [missedSales]
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of missed sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MissedSale"
    post:
      operationId: createMissedSale
      tags: [missedSales]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMissedSaleBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissedSale"
  /missed-sales/{id}:
    delete:
      operationId: deleteMissedSale
      tags: [missedSales]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DELIVERIES ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Medicine:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        genericNameId:
          type: ["integer", "null"]
        genericName:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        unitId:
          type: ["integer", "null"]
        unitName:
          type: ["string", "null"]
        rackId:
          type: ["integer", "null"]
        rackName:
          type: ["string", "null"]
        strength:
          type: ["string", "null"]
        packingLabel:
          type: ["string", "null"]
        conversionFactor:
          type: integer
        salePrice:
          type: number
        purchasePrice:
          type: number
        reorderLevel:
          type: integer
        isControlled:
          type: boolean
        defaultSaleUnit:
          type: string
        barcode:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, conversionFactor, salePrice, purchasePrice, reorderLevel, isControlled, defaultSaleUnit, createdAt]
    MedicineWithStock:
      allOf:
        - $ref: "#/components/schemas/Medicine"
        - type: object
          properties:
            totalUnits:
              type: integer
            totalPacks:
              type: integer
            nearestExpiry:
              type: ["string", "null"]
          required: [totalUnits, totalPacks]
    CreateMedicineBody:
      type: object
      properties:
        name:
          type: string
        genericNameId:
          type: ["integer", "null"]
        categoryId:
          type: ["integer", "null"]
        companyId:
          type: ["integer", "null"]
        unitId:
          type: ["integer", "null"]
        rackId:
          type: ["integer", "null"]
        strength:
          type: ["string", "null"]
        packingLabel:
          type: ["string", "null"]
        conversionFactor:
          type: integer
        salePrice:
          type: number
        purchasePrice:
          type: number
        reorderLevel:
          type: integer
        isControlled:
          type: boolean
        defaultSaleUnit:
          type: string
        barcode:
          type: ["string", "null"]
      required: [name, conversionFactor, salePrice, purchasePrice, reorderLevel, isControlled, defaultSaleUnit]
    UpdateMedicineBody:
      type: object
      properties:
        name:
          type: string
        genericNameId:
          type: ["integer", "null"]
        categoryId:
          type: ["integer", "null"]
        companyId:
          type: ["integer", "null"]
        unitId:
          type: ["integer", "null"]
        rackId:
          type: ["integer", "null"]
        strength:
          type: ["string", "null"]
        packingLabel:
          type: ["string", "null"]
        conversionFactor:
          type: integer
        salePrice:
          type: number
        purchasePrice:
          type: number
        reorderLevel:
          type: integer
        isControlled:
          type: boolean
        defaultSaleUnit:
          type: string
        barcode:
          type: ["string", "null"]
    Batch:
      type: object
      properties:
        id:
          type: integer
        medicineId:
          type: integer
        batchNo:
          type: string
        expiryDate:
          type: string
        quantityUnits:
          type: integer
        purchasePrice:
          type: number
        salePrice:
          type: number
      required: [id, medicineId, batchNo, expiryDate, quantityUnits, purchasePrice, salePrice]
...
    PurchaseReturn:
      type: object
      properties:
        id:
          type: integer
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        date:
          type: string
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, date, totalAmount, createdAt]
    PurchaseReturnItem:
      type: object
      properties:
        id:
          type: integer
        medicineId:
          type: integer
        medicineName:
          type: string
        batchId:
          type: ["integer", "null"]
        batchNo:
          type: ["string", "null"]
        returnPacks:
          type: number
        returnUnits:
          type: number
        purchasePrice:
          type: number
        totalAmount:
          type: number
      required: [id, medicineId, medicineName, returnPacks, returnUnits, purchasePrice, totalAmount]
    PurchaseReturnWithItems:
      allOf:
        - $ref: "#/components/schemas/PurchaseReturn"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/PurchaseReturnItem"
          required: [items]
    CreatePurchaseReturnBody:
      type: object
      properties:
        supplierId:
          type: ["integer", "null"]
        date:
          type: string
        notes:
          type: ["string", "null"]
        inPacks:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreatePurchaseReturnItemBody"
      required: [date, items, inPacks]
    CreatePurchaseReturnItemBody:
      type: object
      properties:
        medicineId:
          type: integer
        batchId:
          type: ["integer", "null"]
        returnQuantity:
          type: number
        purchasePrice:
          type: number
      required: [medicineId, returnQuantity, purchasePrice]

    # ─── SALES ─────────────────────────────────────────────────────────────────
...
    MissedSale:
      type: object
      properties:
        id:
          type: integer
        medicineName:
          type: string
        genericName:
          type: ["string", "null"]
        quantityDemanded:
          type: number
        customerNote:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string
      required: [id, medicineName, quantityDemanded, date, createdAt]
    CreateMissedSaleBody:
      type: object
      properties:
        medicineName:
          type: string
        genericName:
          type: ["string", "null"]
        quantityDemanded:
          type: number
        customerNote:
          type: ["string", "null"]
        date:
          type: string
      required: [medicineName, quantityDemanded, date]

    # ─── DELIVERIES ────────────────────────────────────────────────────────────
...