# lib/api-zod/src/generated/types/missedSale.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PharmaCare Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface MissedSale {
  id: number;
  medicineName: string;
  /** @nullable */
  genericName?: string | null;
  quantityDemanded: number;
  /** @nullable */
  customerNote?: string | null;
  date: string;
  createdAt: string;
}