# lib/api-zod/src/generated/types/purchaseReturn.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PharmaCare Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface PurchaseReturn {
  id: number;
  /** @nullable */
  supplierId?: number | null;
  /** @nullable */
  supplierName?: string | null;
  date: string;
  totalAmount: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}